Req OLTSWITCH-1543 Upgrade SDK version to 6.5.25 on TC16SXG card Req OLTSWITCH-1544 Upgrade SDK version to 6.5.25 on version 3.18 for all eligible cards -Compilation was...  ...working for TC16SXG and CXO160G;  ...failing for CXO640G, TG16GF and OLT1T0F.
svn path=/trunk/; revision=8479

+++ lib/broadcom-sdk-xgs/sdk-all-switchdrvr-6.5.25/broadcom/esw_sdk/systems/bde/linux/kernel/linux_dma.c
@@ -961,7 +961,7 @@ void _dma_per_device_init(int dev_index)
         if (!_dma_vbase || !dma_addr) {
             _dma_pool_alloc_state = DMA_POOL_FAILED;
             gprintk("%s: failed to allocate coherent memory pool of size 0x%x (_dma_vbase=0x%08lx dma_addr=0x%08llx)\n",
-                    __FUNCTION__, _dma_mem_size, (unsigned long) _dma_vbase, dma_addr);
+                    __FUNCTION__, _dma_mem_size, (unsigned long) _dma_vbase, (unsigned long long) dma_addr);
             return;
         }
         _dma_alloc_coherent_device = dev;
